Replace ViewPager2 and task fragments with Compose-based TaskScreenContainer
