--- app/src/main/java/org/groundplatform/android/ui/datacollection/DataCollectionViewPagerAdapterFactory.kt
@@ -1,25 +1,0 @@
-/*
- * Copyright 2022 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.groundplatform.android.ui.datacollection
-
-import androidx.fragment.app.Fragment
-import dagger.assisted.AssistedFactory
-import org.groundplatform.domain.model.task.Task
-
-@AssistedFactory
-interface DataCollectionViewPagerAdapterFactory {
-  fun create(fragment: Fragment, tasks: List<Task>): DataCollectionViewPagerAdapter
-}
